semaphore: Fix a few regressions
Regressions caused by signedness issues in
"sem: change sem wait to atomic operation".
(https://github.com/apache/nuttx/pull/14465)

An alternative would be to make these atomic macros propagate
signedness using the typeof() GCC/clang extension. I'm not inclined
to do so because typeof is not so portable though. As we can unlikely
require "real" C11 atomics in the foreseeable future, maybe we should
use a different set of names from C11 to avoid confusions.

+++ sched/semaphore/sem_holder.c
@@ -880,7 +880,7 @@ void nxsem_canceled(FAR struct tcb_s *stcb, FAR sem_t *sem)
 {
   /* Check our assumptions */
 
-  DEBUGASSERT(atomic_load(NXSEM_COUNT(sem)) <= 0);
+  DEBUGASSERT((int16_t)atomic_load(NXSEM_COUNT(sem)) <= 0);
 
   /* Adjust the priority of every holder as necessary */
